second program, fixes signals
Testing signal to show on oscilloscope that signal b and a can have a delay after each other. The second file is scaled by x100 from v2 as this shows the logic on the oscilloscope better. I fixed the signal length as well as there was a problem that there wasn't a temporary perameter thus the parameter was going up to 1 second instead of  700microseconds